+++ modules/test/single_arch_vmm_test.c
@@ -44,7 +44,7 @@ int arch_vmm_test(void)
         /*=== === === ===*/
 
         /*TEST:memset the map l3 table to 0*/
-        u32 ppn = buddy_pmm.pmm_alloc(1, ZONE_NORMAL);
+        i64 ppn = buddy_pmm.pmm_alloc(1, ZONE_NORMAL);
         paddr page_paddr = PADDR(ppn);
         vaddr page_vaddr = map_pages;
         ENTRY_FLAGS_t flags = arch_decode_flags(
@@ -70,7 +70,7 @@ int arch_vmm_test(void)
         struct vspace *vs = percpu(current_vspace);
         /* actually we should lock this pmm_alloc, but it's a test, we think
          * there's no preemt*/
-        int ppn_1 = buddy_pmm.pmm_alloc(1, ZONE_NORMAL);
+        i64 ppn_1 = buddy_pmm.pmm_alloc(1, ZONE_NORMAL);
         if (ppn_1 <= 0) {
                 pr_error("[ ERROR ] ERROR:try get a ppn fail\n");
                 goto arch_vmm_test_error;
@@ -95,7 +95,7 @@ int arch_vmm_test(void)
                 then we alloc another page frame and try to map to same virtual
            region ,expect fail
         */
-        int ppn_2 = buddy_pmm.pmm_alloc(1, ZONE_NORMAL);
+        i64 ppn_2 = buddy_pmm.pmm_alloc(1, ZONE_NORMAL);
         if (ppn_2 <= 0) {
                 pr_error("[ ERROR ] ERROR:try get a ppn fail\n");
                 goto arch_vmm_test_error;
@@ -115,7 +115,7 @@ int arch_vmm_test(void)
         /*
                 this time alloc a new 2M virtual region
         */
-        int ppn_3 = buddy_pmm.pmm_alloc(MIDDLE_PAGES, ZONE_NORMAL);
+        i64 ppn_3 = buddy_pmm.pmm_alloc(MIDDLE_PAGES, ZONE_NORMAL);
         if (ppn_3 <= 0) { /*we expect the ppn aligned*/
                 pr_error("[ ERROR ] ERROR:try get a ppn fail\n");
                 goto arch_vmm_test_error;
@@ -136,7 +136,7 @@ int arch_vmm_test(void)
                 another 2M page map to same virtual region
                                 except same fail as 4K remap
         */
-        int ppn_4 = buddy_pmm.pmm_alloc(MIDDLE_PAGES, ZONE_NORMAL);
+        i64 ppn_4 = buddy_pmm.pmm_alloc(MIDDLE_PAGES, ZONE_NORMAL);
         if (ppn_4 <= 0) { /*we expect the ppn aligned*/
                 pr_error("[ ERROR ] ERROR:try get a ppn fail\n");
                 goto arch_vmm_test_error;
